refactor: Consolidate deprecated aliases in src/index.ts
