+++ src/feeds/json/common/types.ts
@@ -80,12 +80,3 @@ export namespace JsonFeed {
   >
 }
 // #endregion reference
-
-/** @deprecated Use `JsonFeed` instead. Will be removed in the next major version. */
-export namespace Json {
-  export type Author = JsonFeed.Author
-  export type Attachment<TStrict extends boolean = false> = JsonFeed.Attachment<TStrict>
-  export type Item<TDate, TStrict extends boolean = false> = JsonFeed.Item<TDate, TStrict>
-  export type Hub<TStrict extends boolean = false> = JsonFeed.Hub<TStrict>
-  export type Feed<TDate, TStrict extends boolean = false> = JsonFeed.Feed<TDate, TStrict>
-}
